feat: check for ddos before upload files to fs
